Base Requirements Complete
Debug Results:
By setting "newBoard" equal to "board" in negamax instead of using a
clone, the reference was being copied and changes made to actual board
during negamax.

The move made too early was the same move as it was going to make anyway
so I could've kept it and adjusted for it, but I wanted to keep negamax
to only the move planning and separate the actual move making.